Updating Create-Token to Create-Embed-Token
