+++ bin/cli-create-embed-token
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../lib/cli-create-embed-token.js').default();
