--- bin/cli-create-token
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require('../lib/cli-create-token.js').default();
